backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
httpx==0.27.2
pyjwt==2.9.0
python-multipart==0.0.9

redis==5.0.8